requirements.txt
cachetools==5.5.0
numpy==2.2.1
python-dotenv==1.0.1
web3==7.7.0
